customer_support_agent/services/knowledge_base/requirements.txt
fastapi
uvicorn
httpx
pymupdf
langchain
langchain-ollama
langchain-text-splitters
python-multipart
sqlalchemy
psycopg2-binary
pgvector
streamlit
pyyaml